+++ ui/screens/__init__.py
@@ -1,18 +1,18 @@
 """UI screen modules for WorkoutApp."""
 
 from .exercise_library import ExerciseLibraryScreen
-
-__all__ = ["ExerciseLibraryScreen"]
 from .metric_input_screen import MetricInputScreen
-
-__all__ = ["MetricInputScreen"]
-
-"""Screen module exports."""
-
 from .workout_active_screen import WorkoutActiveScreen
-
-__all__ = ["WorkoutActiveScreen"]
-
 from .preset_detail_screen import PresetDetailScreen
+from .presets_screen import PresetsScreen
+from .workout_summary_screen import WorkoutSummaryScreen
 
+__all__ = [
+    "ExerciseLibraryScreen",
+    "MetricInputScreen",
+    "WorkoutActiveScreen",
+    "PresetDetailScreen",
+    "PresetsScreen",
+    "WorkoutSummaryScreen",
+]
