+++ ui/screens/workout_summary_screen.py
@@ -1,0 +1,32 @@
+from kivymd.app import MDApp
+from kivymd.uix.screen import MDScreen
+from kivymd.uix.list import OneLineListItem
+from kivy.properties import ObjectProperty
+
+
+class WorkoutSummaryScreen(MDScreen):
+    """Screen showing the summary of a completed workout."""
+
+    summary_list = ObjectProperty(None)
+
+    def on_pre_enter(self, *args):
+        self.populate()
+        return super().on_pre_enter(*args)
+
+    def populate(self):
+        if not self.summary_list:
+            return
+        self.summary_list.clear_widgets()
+        app = MDApp.get_running_app()
+        session = app.workout_session
+        if not session:
+            return
+        print(session.summary())
+        for exercise in session.exercises:
+            self.summary_list.add_widget(OneLineListItem(text=exercise["name"]))
+            for idx, metrics in enumerate(exercise["results"], 1):
+                metrics_text = ", ".join(f"{k}: {v}" for k, v in metrics.items())
+                self.summary_list.add_widget(
+                    OneLineListItem(text=f"Set {idx}: {metrics_text}")
+                )
+
